Proxy domain
When testing the proxy over OpenVPN, I found that iOS would send an SOA
question for the .local TLD when attempting to resolve addresses.

To get around this, I have set up the app to use a custom TLD (.hippo)
and swap requests for it with .local address name resolution locally.
This isn't ideal, as it means that hosts inside the network need to be
aware of the proxy DNS if I want to use those address transparently,
but it will work for now.